Spawn thread-safe actors in spawn module
